+++ frontend/src/components/dialogs/AssetDetailDialog.vue
@@ -1,39 +1,7 @@
 <template>
     <ff-dialog ref="dialog" :header="header" confirm-label="Close" :closeOnConfirm="true" data-el="flow-view-dialog" boxClass="!min-w-[80%] !min-h-[80%] !w-[80%] !h-[80%]" contentClass="overflow-hidden flex-grow" @confirm="confirm()">
         <template #default>
-            <div v-if="mode==='compare'" class="w-full h-full">
-                <div class="flex gap-2" data-el="snapshot-compare-toolbar">
-                    <ff-dropdown v-model="compareSnapshot" class="flex-grow">
-                        <ff-dropdown-option v-for="option in compareSnapshotList" :key="option.value" :value="option.value" :label="option.label" :title="option.description" class="text-sm">
-                            {{ option.label }}
-                        </ff-dropdown-option>
-                    </ff-dropdown>
-                    <ff-button
-                        v-if="true"
-                        :disabled="!compareSnapshot"
-                        data-action="compare-snapshots"
-                        kind="secondary"
-                        style="height: 30px; width: 106px"
-                        class="w-32"
-                        @click="renderComparison"
-                    >
-                        Compare
-                    </ff-button>
-                </div>
-                <div v-if="changes.length" class="flex justify-between items-center gap-2 mt-2 ml-2">
-                    <div class="whitespace-nowrap">Change {{ changeIndex + 1 }} of {{ changes.length }}:</div>
-                    <div class="text-sm text-gray-500 flex-grow truncate overflow-ellipsis">{{ changes[changeIndex].toString() }}</div>
-                    <ff-button kind="secondary" size="small" class="w-14" @click="gotoPreviousDifference">Prev</ff-button>
-                    <ff-button kind="secondary" size="small" class="w-14" @click="gotoNextDifference">Next</ff-button>
-                </div>
-                <div v-else class="mt-2">
-                    <div class="text-sm text-gray-500 flex-grow truncate overflow-ellipsis ml-2">No differences found</div>
-                </div>
-                <div ref="compareViewer" data-el="ff-flow-compare-view" class="ff-flow-compare-viewer pt-4" @click.stop.prevent>
-                    &nbsp;
-                </div>
-            </div>
-            <div v-else ref="viewer" data-el="ff-flow-previewer" class="ff-flow-viewer" @click.stop.prevent>
+            <div ref="viewer" data-el="ff-flow-previewer" class="ff-flow-viewer" @click.stop.prevent>
                 Loading...
             </div>
         </template>
@@ -47,8 +15,6 @@
 <script>
 
 import FlowRenderer from '@flowfuse/flow-renderer'
-
-import Alerts from '../../services/alerts.js'
 
 export default {
     name: 'FlowViewerDialog',
@@ -67,35 +33,12 @@ export default {
                 setTimeout(() => {
                     this.renderFlows()
                 }, 20)
-            },
-            /**
-             * Shows the compare flows dialog and presents the user with a list of snapshots to compare against
-             * @param {{flows: { flows :[]}}} v1Snapshot - A snapshot object as the base for comparison
-             * @param {[{label: String, value: String}]} snapshotList - A list of snapshots to compare against where label is the snapshot name and value is the snapshot id
-             * @param {(snapshotId)=>Promise<{flows: { flows :[]}}>} snapshotLoaderCallback - A callback function that accepts a snapshot id and returns the snapshot object
-             */
-            showCompareSnapshots (v1Snapshot, snapshotList, snapshotLoaderCallback) {
-                this.mode = 'compare'
-                this.payload = v1Snapshot
-                this.compareSnapshot = null
-                this.changes = []
-                this.changeIndex = 0
-                this.compareSnapshotList = snapshotList
-                this.getSnapshotCallback = snapshotLoaderCallback
-                this.$refs.dialog.show()
             }
         }
     },
     data () {
         return {
-            payload: [],
-            snapshotList: [],
-            compareSnapshot: null,
-            compareSnapshotList: [],
-            mode: 'view', // view, compare
-            getSnapshotCallback: null,
-            changes: [],
-            changeIndex: 0
+            payload: []
         }
     },
     computed: {
@@ -110,45 +53,13 @@ export default {
     },
     methods: {
         confirm () {
-            this.cleanup()
             this.$refs.dialog.close()
         },
         renderFlows () {
-            this.cleanup()
             const flowRenderer = new FlowRenderer()
             flowRenderer.renderFlows(this.flow, {
                 container: this.$refs.viewer
             })
-        },
-        async renderComparison (snapshotId) {
-            this.cleanup()
-            const compareSnapshot = await this.getSnapshotCallback(this.compareSnapshot)
-            if (!compareSnapshot?.flows?.flows) {
-                Alerts.emit('Flows not found in the selected snapshot', 'warning')
-                return
-            }
-            const flowRenderer = new FlowRenderer()
-            const flows = [this.flow, compareSnapshot?.flows?.flows]
-            const result = flowRenderer.compare(flows, {
-                container: this.$refs.compareViewer
-            })
-            this.changes = result?.changes || []
-        },
-        gotoNextDifference () {
-            this.changeIndex = (this.changeIndex + 1) % this.changes.length
-            this.changes[this.changeIndex].highlight()
-        },
-        gotoPreviousDifference () {
-            this.changeIndex = (this.changeIndex - 1 + this.changes.length) % this.changes.length
-            this.changes[this.changeIndex].highlight()
-        },
-        cleanup () {
-            while (this.$refs.compareViewer?.firstChild) {
-                this.$refs.compareViewer.removeChild(this.$refs.compareViewer.firstChild)
-            }
-            while (this.$refs.viewer?.firstChild) {
-                this.$refs.viewer.removeChild(this.$refs.viewer.firstChild)
-            }
         }
     }
 }
@@ -157,8 +68,5 @@ export default {
 <style scoped>
 .ff-flow-viewer {
     height: 100%;
-}
-.ff-flow-compare-viewer {
-    height: calc(100% - 4.5rem);
 }
 </style>
