+++ frontend/src/pages/instance/Snapshots/index.vue
@@ -58,7 +58,7 @@
         <SnapshotExportDialog ref="snapshotExportDialog" data-el="dialog-export-snapshot" :project="instance" />
         <SnapshotImportDialog ref="snapshotImportDialog" title="Upload Snapshot" data-el="dialog-import-snapshot" :owner="instance" owner-type="instance" @snapshot-import-success="onSnapshotImportSuccess" @snapshot-import-failed="onSnapshotImportFailed" @canceled="onSnapshotImportCancel" />
         <AssetDetailDialog ref="snapshotViewerDialog" data-el="dialog-view-snapshot" />
-        <AssetDetailDialog ref="snapshotCompareDialog" data-el="dialog-compare-snapshot" />
+        <AssetCompareDialog ref="snapshotCompareDialog" data-el="dialog-compare-snapshot" />
     </div>
 </template>
 
@@ -73,6 +73,7 @@ import SnapshotsApi from '../../../api/snapshots.js'
 
 import EmptyState from '../../../components/EmptyState.vue'
 import SectionTopMenu from '../../../components/SectionTopMenu.vue'
+import AssetCompareDialog from '../../../components/dialogs/AssetCompareDialog.vue'
 import AssetDetailDialog from '../../../components/dialogs/AssetDetailDialog.vue'
 import SnapshotImportDialog from '../../../components/dialogs/SnapshotImportDialog.vue'
 import UserCell from '../../../components/tables/cells/UserCell.vue'
@@ -98,7 +99,8 @@ export default {
         SnapshotImportDialog,
         PlusSmIcon,
         UploadIcon,
-        AssetDetailDialog
+        AssetDetailDialog,
+        AssetCompareDialog
     },
     mixins: [permissionsMixin],
     inheritAttrs: false,
@@ -158,6 +160,15 @@ export default {
         },
         busy () {
             return this.busyMakingSnapshot || this.busyImportingSnapshot
+        },
+        snapshotList () {
+            return this.snapshots.map(s => {
+                return {
+                    label: s.name,
+                    description: s.description || '',
+                    value: s.id
+                }
+            })
         }
     },
     watch: {
@@ -270,22 +281,12 @@ export default {
             })
         },
         showCompareSnapshotDialog (snapshot) {
-            SnapshotsApi.getFullSnapshot(snapshot.id).then((data) => {
-                const snapshotList = this.snapshots.map(s => {
-                    return {
-                        label: s.name,
-                        description: s.description || '',
-                        value: s.id
-                    }
+            SnapshotsApi.getFullSnapshot(snapshot.id)
+                .then((data) => this.$refs.snapshotCompareDialog.showCompareSnapshots(data, this.snapshotList))
+                .catch(err => {
+                    console.error(err)
+                    Alerts.emit('Failed to get snapshot.', 'warning')
                 })
-                const snapshotLoaderCallback = (snapshotId) => {
-                    return SnapshotsApi.getFullSnapshot(snapshotId)
-                }
-                this.$refs.snapshotCompareDialog.showCompareSnapshots(data, snapshotList, snapshotLoaderCallback)
-            }).catch(err => {
-                console.error(err)
-                Alerts.emit('Failed to get snapshot.', 'warning')
-            })
         },
         // snapshot actions - import
         showImportSnapshotDialog () {
